+++ motor/coordinator/scheduler/runtime/scheduler_client.py
@@ -199,7 +199,7 @@ class _SchedulerInstanceCache:
         role: PDRole,
         active_tokens: float,
         active_kv_cache: float,
-        prefill_cost: int = 0,
+        prefill_cost: float = 0,
     ) -> None:
         """Patch single endpoint workload from shared memory. Skip if not in cache."""
         role_map = self._instance_map.get(role) or {}
